pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.4",
    "cryptography>=45.0.6",
    "fastapi>=0.116.1",
    "loguru>=0.7.3",
    "notion-client>=2.4.0",
    "openai>=1.99.3",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "requests>=2.32.4",
    "rq>=2.4.1",
    "schedule>=1.2.2",
    "supabase>=2.17.0",
    "tenacity>=9.1.2",
    "uvicorn>=0.35.0",
]
